bootstrap: some code re-org for more clarity
