+++ go/provisioner/local/local.go
@@ -16,115 +16,23 @@ package local
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"net"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"regexp"
 	"slices"
 	"strings"
-	"syscall"
 	"time"
 
 	"github.com/multigres/multigres/go/clustermetadata/topo"
 	"github.com/multigres/multigres/go/provisioner"
-	"github.com/multigres/multigres/go/tools/semver"
 	"github.com/multigres/multigres/go/tools/stringutil"
 
 	clustermetadatapb "github.com/multigres/multigres/go/pb/clustermetadata"
 
 	"gopkg.in/yaml.v3"
 )
-
-// CellConfig holds the configuration for a single cell
-type CellConfig struct {
-	Name     string `yaml:"name"`
-	RootPath string `yaml:"root-path"`
-}
-
-// TopologyConfig holds the configuration for cluster topology
-type TopologyConfig struct {
-	Backend        string       `yaml:"backend"`
-	GlobalRootPath string       `yaml:"global-root-path"`
-	Cells          []CellConfig `yaml:"cells"`
-}
-
-// CellServicesConfig holds the service configuration for a specific cell
-type CellServicesConfig struct {
-	Multigateway MultigatewayConfig `yaml:"multigateway"`
-	Multipooler  MultipoolerConfig  `yaml:"multipooler"`
-	Multiorch    MultiorchConfig    `yaml:"multiorch"`
-	Pgctld       PgctldConfig       `yaml:"pgctld"`
-}
-
-// LocalProvisionerConfig represents the typed configuration for the local provisioner
-type LocalProvisionerConfig struct {
-	RootWorkingDir string                        `yaml:"root-working-dir"`
-	DefaultDbName  string                        `yaml:"default-db-name"`
-	Etcd           EtcdConfig                    `yaml:"etcd"`
-	Topology       TopologyConfig                `yaml:"topology"`
-	Multiadmin     MultiadminConfig              `yaml:"multiadmin"`
-	Cells          map[string]CellServicesConfig `yaml:"cells,omitempty"`
-}
-
-// EtcdConfig holds etcd service configuration
-type EtcdConfig struct {
-	Version string `yaml:"version"`
-	DataDir string `yaml:"data-dir"`
-	Port    int    `yaml:"port"`
-}
-
-// MultigatewayConfig holds multigateway service configuration
-type MultigatewayConfig struct {
-	Path     string `yaml:"path"`
-	HttpPort int    `yaml:"http-port"`
-	GrpcPort int    `yaml:"grpc-port"`
-	PgPort   int    `yaml:"pg-port"`
-	LogLevel string `yaml:"log-level"`
-}
-
-// MultipoolerConfig holds multipooler service configuration
-type MultipoolerConfig struct {
-	Path       string `yaml:"path"`
-	Database   string `yaml:"database"`
-	TableGroup string `yaml:"table-group"`
-	ServiceID  string `yaml:"service-id"`
-	HttpPort   int    `yaml:"http-port"`
-	GrpcPort   int    `yaml:"grpc-port"`
-	LogLevel   string `yaml:"log-level"`
-}
-
-// MultiorchConfig holds multiorch service configuration
-type MultiorchConfig struct {
-	Path     string `yaml:"path"`
-	HttpPort int    `yaml:"http-port"`
-	GrpcPort int    `yaml:"grpc-port"`
-	LogLevel string `yaml:"log-level"`
-}
-
-// MultiadminConfig holds multiadmin service configuration
-type MultiadminConfig struct {
-	Path     string `yaml:"path"`
-	HttpPort int    `yaml:"http-port"`
-	GrpcPort int    `yaml:"grpc-port"`
-	LogLevel string `yaml:"log-level"`
-}
-
-// PgctldConfig holds pgctld service configuration
-type PgctldConfig struct {
-	Path       string `yaml:"path"`
-	PoolerDir  string `yaml:"pooler-dir"`  // Base directory for this pgctld instance
-	GrpcPort   int    `yaml:"grpc-port"`   // gRPC port for pgctld server
-	PgPort     int    `yaml:"pg-port"`     // PostgreSQL port
-	PgDatabase string `yaml:"pg-database"` // PostgreSQL database name
-	PgUser     string `yaml:"pg-user"`     // PostgreSQL username
-	PgPwfile   string `yaml:"pg-pwfile"`   // PostgreSQL password file path (optional)
-	Timeout    int    `yaml:"timeout"`     // Operation timeout in seconds
-	LogLevel   string `yaml:"log-level"`   // Log level
-}
 
 // localProvisioner implements the Provisioner interface for local binary-based provisioning
 type localProvisioner struct {
@@ -135,21 +43,6 @@ type localProvisioner struct {
 // Compile-time check to ensure localProvisioner implements Provisioner
 var _ provisioner.Provisioner = (*localProvisioner)(nil)
 
-// LocalProvisionedService represents a service instance that has been provisioned
-type LocalProvisionedService struct {
-	ID         string         `json:"id"`                    // Unique instance ID
-	Service    string         `json:"service"`               // Service name (etcd, multigateway, etc.)
-	PID        int            `json:"pid,omitempty"`         // For binary processes
-	BinaryPath string         `json:"binary-path,omitempty"` // Path to the binary
-	DataDir    string         `json:"data-dir,omitempty"`    // Data directory
-	LogFile    string         `json:"log-file,omitempty"`    // Path to log file
-	Ports      map[string]int `json:"ports"`                 // Port mappings
-	FQDN       string         `json:"fqdn"`                  // Hostname/FQDN
-	Runtime    string         `json:"runtime"`               // "binary"
-	StartedAt  time.Time      `json:"started-at"`            // When it was started
-	Metadata   map[string]any `json:"metadata,omitempty"`    // Additional metadata
-}
-
 const (
 	// StateDir is the directory name where provision state files are stored
 	StateDir = "state"
@@ -158,194 +51,6 @@ const (
 // Name returns the name of this provisioner
 func (p *localProvisioner) Name() string {
 	return "local"
-}
-
-// LoadConfig loads the provisioner-specific configuration from the given config paths
-func (p *localProvisioner) LoadConfig(configPaths []string) error {
-	// Try to find the config file in the provided paths
-	for _, configPath := range configPaths {
-		configFile := filepath.Join(configPath, "multigres.yaml")
-		if _, err := os.Stat(configFile); err == nil {
-			data, err := os.ReadFile(configFile)
-			if err != nil {
-				return fmt.Errorf("failed to read config file %s: %w", configFile, err)
-			}
-
-			// Parse the full config file
-			var fullConfig struct {
-				Provisioner       string         `yaml:"provisioner"`
-				ProvisionerConfig map[string]any `yaml:"provisioner-config,omitempty"`
-			}
-			if err := yaml.Unmarshal(data, &fullConfig); err != nil {
-				return fmt.Errorf("failed to parse config file %s: %w", configFile, err)
-			}
-
-			// Validate that this is for the local provisioner
-			if fullConfig.Provisioner != "local" {
-				return fmt.Errorf("config file %s is for provisioner '%s', not 'local'", configFile, fullConfig.Provisioner)
-			}
-
-			if err := p.ValidateConfig(fullConfig.ProvisionerConfig); err != nil {
-				return fmt.Errorf("failed to validate config file %s: %w", configFile, err)
-			}
-
-			// Convert the provisioner-config section to our typed config
-			yamlData, err := yaml.Marshal(fullConfig.ProvisionerConfig)
-			if err != nil {
-				return fmt.Errorf("failed to marshal provisioner config: %w", err)
-			}
-
-			p.config = &LocalProvisionerConfig{}
-			if err := yaml.Unmarshal(yamlData, p.config); err != nil {
-				return fmt.Errorf("failed to unmarshal provisioner config: %w", err)
-			}
-
-			return nil
-		}
-	}
-
-	return fmt.Errorf("multigres.yaml not found in any of the provided paths: %v", configPaths)
-}
-
-// DefaultConfig returns the default configuration for the local provisioner
-func (p *localProvisioner) DefaultConfig() map[string]any {
-	// Use MTROOT environment variable if set, otherwise fall back to current directory
-	mtroot := os.Getenv("MTROOT")
-	baseDir := "."
-	binDir := "bin"
-
-	if mtroot != "" {
-		baseDir = mtroot + "/multigres_local"
-		binDir = filepath.Join(mtroot, "bin")
-	} else {
-		fmt.Println("Warning: MTROOT environment variable is not set, using relative paths for default binary configuration in local provisioner.")
-	}
-
-	// Generate service IDs for each cell using the same method as topo components
-	serviceIDZone1 := stringutil.RandomString(8)
-	serviceIDZone2 := stringutil.RandomString(8)
-	tableGroup := "default"
-	dbName := "postgres"
-
-	// Create typed configuration with defaults
-	localConfig := LocalProvisionerConfig{
-		RootWorkingDir: baseDir,
-		DefaultDbName:  dbName,
-		Etcd: EtcdConfig{
-			Version: "3.5.9",
-			DataDir: filepath.Join(baseDir, "data", "etcd-data"),
-			Port:    2379,
-		},
-		Topology: TopologyConfig{
-			Backend:        "etcd2",
-			GlobalRootPath: "/multigres/global",
-			Cells: []CellConfig{
-				{
-					Name:     "zone1",
-					RootPath: "/multigres/zone1",
-				},
-				{
-					Name:     "zone2",
-					RootPath: "/multigres/zone2",
-				},
-			},
-		},
-		Multiadmin: MultiadminConfig{
-			Path:     filepath.Join(binDir, "multiadmin"),
-			HttpPort: 15000,
-			GrpcPort: 15990,
-			LogLevel: "info",
-		},
-		Cells: map[string]CellServicesConfig{
-			"zone1": {
-				Multigateway: MultigatewayConfig{
-					Path:     filepath.Join(binDir, "multigateway"),
-					HttpPort: 15001,
-					GrpcPort: 15991,
-					PgPort:   15432,
-					LogLevel: "info",
-				},
-				Multipooler: MultipoolerConfig{
-					Path:       filepath.Join(binDir, "multipooler"),
-					Database:   dbName,
-					TableGroup: tableGroup,
-					ServiceID:  serviceIDZone1,
-					HttpPort:   15100,
-					GrpcPort:   16001,
-					LogLevel:   "info",
-				},
-				Multiorch: MultiorchConfig{
-					Path:     filepath.Join(binDir, "multiorch"),
-					HttpPort: 15300,
-					GrpcPort: 16000,
-					LogLevel: "info",
-				},
-				Pgctld: PgctldConfig{
-					Path:       filepath.Join(binDir, "pgctld"),
-					PoolerDir:  GeneratePoolerDir(baseDir, serviceIDZone1),
-					GrpcPort:   17000,
-					PgPort:     5432,
-					PgDatabase: dbName,
-					PgUser:     "postgres",
-					PgPwfile:   filepath.Join(GeneratePoolerDir(baseDir, serviceIDZone1), "pgpassword.txt"),
-					Timeout:    30,
-					LogLevel:   "info",
-				},
-			},
-			"zone2": {
-				Multigateway: MultigatewayConfig{
-					Path:     filepath.Join(binDir, "multigateway"),
-					HttpPort: 15101, // zone1 + 100
-					GrpcPort: 16091, // zone1 + 100
-					PgPort:   15532, // zone1 + 100
-					LogLevel: "info",
-				},
-				Multipooler: MultipoolerConfig{
-					Path:       filepath.Join(binDir, "multipooler"),
-					Database:   dbName,
-					TableGroup: tableGroup,
-					ServiceID:  serviceIDZone2,
-					HttpPort:   15200, // zone1 + 100
-					GrpcPort:   16101, // zone1 + 100
-					LogLevel:   "info",
-				},
-				Multiorch: MultiorchConfig{
-					Path:     filepath.Join(binDir, "multiorch"),
-					HttpPort: 15400, // zone1 + 100
-					GrpcPort: 16100, // zone1 + 100
-					LogLevel: "info",
-				},
-				Pgctld: PgctldConfig{
-					Path:       filepath.Join(binDir, "pgctld"),
-					PoolerDir:  GeneratePoolerDir(baseDir, serviceIDZone2),
-					GrpcPort:   17100, // zone1 + 100
-					PgPort:     5532,  // zone1 + 100
-					PgDatabase: dbName,
-					PgUser:     "postgres",
-					PgPwfile:   filepath.Join(GeneratePoolerDir(baseDir, serviceIDZone2), "pgpassword.txt"),
-					Timeout:    30,
-					LogLevel:   "info",
-				},
-			},
-		},
-	}
-
-	// Convert to map[string]any via YAML marshaling to preserve struct ordering
-	yamlData, err := yaml.Marshal(localConfig)
-	if err != nil {
-		// Fallback to empty config if marshaling fails
-		fmt.Printf("Warning: failed to marshal default config: %v\n", err)
-		return map[string]any{}
-	}
-
-	var configMap map[string]any
-	if err := yaml.Unmarshal(yamlData, &configMap); err != nil {
-		// Fallback to empty config if unmarshaling fails
-		fmt.Printf("Warning: failed to unmarshal default config: %v\n", err)
-		return map[string]any{}
-	}
-
-	return configMap
 }
 
 // createPasswordFileAndDirectories creates the pooler directory structure and password file
@@ -535,125 +240,6 @@ func (p *localProvisioner) provisionEtcd(ctx context.Context, req *provisioner.P
 	}, nil
 }
 
-// findBinary finds a binary by name, checking PATH first, then optional configured path
-func (p *localProvisioner) findBinary(name string, serviceConfig map[string]any) (string, error) {
-	// First try to find in PATH
-	if binaryPath, err := exec.LookPath(name); err == nil {
-		return binaryPath, nil
-	}
-
-	// Then try configured path if provided
-	if pathConfig, ok := serviceConfig["path"].(string); ok && pathConfig != "" {
-		// Check if it's an absolute path or relative path
-		var fullPath string
-		if filepath.IsAbs(pathConfig) {
-			fullPath = pathConfig
-		} else {
-			// Make it relative to current directory
-			fullPath = filepath.Join(".", pathConfig)
-		}
-
-		// Check if the binary exists and is executable
-		if info, err := os.Stat(fullPath); err == nil && !info.IsDir() {
-			return fullPath, nil
-		}
-	}
-
-	return "", fmt.Errorf("binary '%s' not found in PATH or configured path", name)
-}
-
-// checkEtcdVersion verifies that the etcd binary major version matches expected version
-func (p *localProvisioner) checkEtcdVersion(binaryPath, expectedVersion string) error {
-	// Run etcd --version to get version info
-	cmd := exec.Command(binaryPath, "--version")
-	output, err := cmd.Output()
-	if err != nil {
-		return fmt.Errorf("failed to get etcd version: %w", err)
-	}
-
-	// Parse version from output - etcd version output format varies
-	versionStr := string(output)
-
-	// Try to extract version number from various etcd output formats
-	versionRegex := regexp.MustCompile(`(?:etcd\s+Version:\s*|^|\s+)v?(\d+\.\d+\.\d+)`)
-	matches := versionRegex.FindStringSubmatch(versionStr)
-	if len(matches) < 2 {
-		// If we can't parse version, just warn and continue
-		fmt.Printf("Warning: could not parse etcd version from output: %s\n", strings.TrimSpace(versionStr))
-		return nil
-	}
-
-	actualVersion := "v" + matches[1] // ensure v prefix for semver
-	expectedVersionWithV := "v" + strings.TrimPrefix(expectedVersion, "v")
-
-	// Use servenv semver to compare major versions
-	actualMajor := semver.Major(actualVersion)
-	expectedMajor := semver.Major(expectedVersionWithV)
-
-	if actualMajor != expectedMajor {
-		return fmt.Errorf("etcd major version mismatch: expected %s.x.x, found %s",
-			strings.TrimPrefix(expectedMajor, "v"), strings.TrimPrefix(actualVersion, "v"))
-	}
-
-	fmt.Printf("🔍 - etcd %s found — version compatible ✓\n",
-		strings.TrimPrefix(actualVersion, "v"))
-	return nil
-}
-
-// readServiceLogs reads the last few lines from a service's log file for debugging
-func (p *localProvisioner) readServiceLogs(logFile string, lines int) string {
-	if logFile == "" {
-		return "No log file available"
-	}
-
-	// Check if log file exists
-	if _, err := os.Stat(logFile); os.IsNotExist(err) {
-		return fmt.Sprintf("Log file not found: %s", logFile)
-	}
-
-	// Read the file
-	data, err := os.ReadFile(logFile)
-	if err != nil {
-		return fmt.Sprintf("Failed to read log file %s: %v", logFile, err)
-	}
-
-	// Get the last N lines
-	logLines := strings.Split(strings.TrimSpace(string(data)), "\n")
-	if len(logLines) == 0 {
-		return "Log file is empty"
-	}
-
-	// Return last 'lines' lines or all lines if fewer exist
-	start := max(len(logLines)-lines, 0)
-
-	result := strings.Join(logLines[start:], "\n")
-	if result == "" {
-		return "Log file is empty"
-	}
-
-	return result
-}
-
-// validateProcessRunning checks if a process with the given PID is still running
-func (p *localProvisioner) validateProcessRunning(pid int) error {
-	if pid <= 0 {
-		return fmt.Errorf("invalid PID: %d", pid)
-	}
-
-	process, err := os.FindProcess(pid)
-	if err != nil {
-		return fmt.Errorf("process with PID %d not found: %w", pid, err)
-	}
-
-	// Send signal 0 to check if process exists without actually sending a signal
-	err = process.Signal(syscall.Signal(0))
-	if err != nil {
-		return fmt.Errorf("process with PID %d is not running: %w", pid, err)
-	}
-
-	return nil
-}
-
 // getRootWorkingDir returns the root working directory from config
 func (p *localProvisioner) getRootWorkingDir() string {
 	if p.config == nil {
@@ -661,25 +247,6 @@ func (p *localProvisioner) getRootWorkingDir() string {
 	}
 
 	return p.config.RootWorkingDir
-}
-
-// GeneratePoolerDir generates a pooler directory path for a given base directory and service ID
-func GeneratePoolerDir(baseDir, serviceID string) string {
-	return filepath.Join(baseDir, "data", fmt.Sprintf("pooler_%s", serviceID))
-}
-
-// getStateDir returns the path to the state directory
-func (p *localProvisioner) getStateDir() string {
-	return filepath.Join(p.getRootWorkingDir(), "state")
-}
-
-// getLogsDir returns the path to the logs directory
-func (p *localProvisioner) getLogsDir() string {
-	return filepath.Join(p.getRootWorkingDir(), "logs")
-}
-
-func (p *localProvisioner) getDataDir() string {
-	return filepath.Join(p.getRootWorkingDir(), "data")
 }
 
 // createLogFile creates a log file path and ensures the directory exists
@@ -722,62 +289,6 @@ func (p *localProvisioner) cleanupLogFile(logFilePath string) error {
 	}
 
 	fmt.Printf("Cleaned up log file: %s\n", logFilePath)
-	return nil
-}
-
-// saveServiceState saves the provisioned service state to disk
-func (p *localProvisioner) saveServiceState(service *LocalProvisionedService, databaseName string) error {
-	stateDir := p.getStateDir()
-	var targetDir string
-
-	if databaseName != "" {
-		// For database services: state/dbs/dbname
-		targetDir = filepath.Join(stateDir, "dbs", databaseName)
-	} else {
-		// For non-database services (like etcd): state/
-		targetDir = stateDir
-	}
-
-	if err := os.MkdirAll(targetDir, 0o755); err != nil {
-		return fmt.Errorf("failed to create state directory %s: %w", targetDir, err)
-	}
-
-	// File name format: service_id.json (e.g., etcd_abc123.json)
-	fileName := fmt.Sprintf("%s_%s.json", service.Service, service.ID)
-	filePath := filepath.Join(targetDir, fileName)
-
-	data, err := json.MarshalIndent(service, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal service state: %w", err)
-	}
-
-	if err := os.WriteFile(filePath, data, 0o644); err != nil {
-		return fmt.Errorf("failed to write service state file %s: %w", filePath, err)
-	}
-
-	return nil
-}
-
-// removeServiceState removes a service state file from disk
-func (p *localProvisioner) removeServiceState(serviceID, serviceName, databaseName string) error {
-	stateDir := p.getStateDir()
-	var targetDir string
-
-	if databaseName != "" {
-		// For database services: state/dbs/dbname
-		targetDir = filepath.Join(stateDir, "dbs", databaseName)
-	} else {
-		// For non-database services (like etcd): state/
-		targetDir = stateDir
-	}
-
-	fileName := fmt.Sprintf("%s_%s.json", serviceName, serviceID)
-	filePath := filepath.Join(targetDir, fileName)
-
-	if err := os.Remove(filePath); err != nil && !os.IsNotExist(err) {
-		return fmt.Errorf("failed to remove state file %s: %w", filePath, err)
-	}
-
 	return nil
 }
 
@@ -1385,388 +896,6 @@ func (p *localProvisioner) Deprovision(ctx context.Context, req *provisioner.Dep
 	return nil
 }
 
-// loadServiceState loads a specific service state from disk
-func (p *localProvisioner) loadServiceState(req *provisioner.DeprovisionRequest) (*LocalProvisionedService, error) {
-	stateDir := p.getStateDir()
-	var targetDir string
-
-	if req.DatabaseName != "" {
-		// For database services: state/dbs/dbname
-		targetDir = filepath.Join(stateDir, "dbs", req.DatabaseName)
-	} else {
-		// For non-database services (like etcd): state/
-		targetDir = stateDir
-	}
-
-	fileName := fmt.Sprintf("%s_%s.json", req.Service, req.ServiceID)
-	filePath := filepath.Join(targetDir, fileName)
-
-	// Check if state file exists
-	if _, err := os.Stat(filePath); os.IsNotExist(err) {
-		return nil, nil // Service not found
-	}
-
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read state file %s: %w", filePath, err)
-	}
-
-	var service LocalProvisionedService
-	if err := json.Unmarshal(data, &service); err != nil {
-		return nil, fmt.Errorf("failed to parse state file %s: %w", filePath, err)
-	}
-
-	// Sanity check: ensure this method is called for the expected service type
-	if req.Service != service.Service {
-		return nil, fmt.Errorf("deprovision%s called for wrong service type: %s", service.Service, req.Service)
-	}
-
-	return &service, nil
-}
-
-// loadDbProvisionedServices loads provisioned services for a specific database
-func (p *localProvisioner) loadDbProvisionedServices(databaseName string) ([]*LocalProvisionedService, error) {
-	if databaseName == "" {
-		return nil, fmt.Errorf("database name is required")
-	}
-
-	stateDir := p.getStateDir()
-	targetDir := filepath.Join(stateDir, "dbs", databaseName)
-
-	// Check if target directory exists
-	if _, err := os.Stat(targetDir); os.IsNotExist(err) {
-		return nil, nil // No services for this database
-	}
-
-	entries, err := os.ReadDir(targetDir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read directory %s: %w", targetDir, err)
-	}
-
-	var services []*LocalProvisionedService
-	for _, entry := range entries {
-		if !entry.IsDir() && strings.HasSuffix(entry.Name(), ".json") {
-			// Parse filename: service_id.json
-			name := strings.TrimSuffix(entry.Name(), ".json")
-			parts := strings.SplitN(name, "_", 2)
-			if len(parts) != 2 {
-				// Skip invalid state files
-				continue
-			}
-
-			serviceName := parts[0]
-			serviceID := parts[1]
-
-			req := &provisioner.DeprovisionRequest{
-				Service:      serviceName,
-				ServiceID:    serviceID,
-				DatabaseName: databaseName,
-			}
-			service, err := p.loadServiceState(req)
-			if err != nil {
-				// Log warning but continue with other services
-				fmt.Printf("Warning: failed to load state for %s service %s: %v\n", serviceName, serviceID, err)
-				continue
-			}
-
-			if service != nil {
-				services = append(services, service)
-			}
-		}
-	}
-
-	return services, nil
-}
-
-// loadEtcdServices loads etcd services from the top-level state directory
-func (p *localProvisioner) loadEtcdServices() ([]*LocalProvisionedService, error) {
-	stateDir := p.getStateDir()
-
-	// Check if state directory exists
-	if _, err := os.Stat(stateDir); os.IsNotExist(err) {
-		return nil, nil // No state directory, no services running
-	}
-
-	entries, err := os.ReadDir(stateDir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read directory %s: %w", stateDir, err)
-	}
-
-	var services []*LocalProvisionedService
-	for _, entry := range entries {
-		if !entry.IsDir() && strings.HasSuffix(entry.Name(), ".json") && strings.HasPrefix(entry.Name(), "etcd_") {
-			// Parse filename: etcd_serviceID.json
-			name := strings.TrimSuffix(entry.Name(), ".json")
-			parts := strings.SplitN(name, "_", 2)
-			if len(parts) != 2 || parts[0] != "etcd" {
-				continue
-			}
-
-			serviceID := parts[1]
-
-			req := &provisioner.DeprovisionRequest{
-				Service:      "etcd",
-				ServiceID:    serviceID,
-				DatabaseName: "", // etcd is a global service
-			}
-			service, err := p.loadServiceState(req)
-			if err != nil {
-				// Log warning but continue with other services
-				fmt.Printf("Warning: failed to load state for etcd service %s: %v\n", serviceID, err)
-				continue
-			}
-
-			if service != nil {
-				services = append(services, service)
-			}
-		}
-	}
-
-	return services, nil
-}
-
-// loadGlobalServices loads all global services (non-database services) from state files
-func (p *localProvisioner) loadGlobalServices() ([]*LocalProvisionedService, error) {
-	stateDir := p.getStateDir()
-
-	// Check if state directory exists
-	if _, err := os.Stat(stateDir); os.IsNotExist(err) {
-		return nil, nil // No state directory, no services running
-	}
-
-	entries, err := os.ReadDir(stateDir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read directory %s: %w", stateDir, err)
-	}
-
-	var services []*LocalProvisionedService
-	for _, entry := range entries {
-		if !entry.IsDir() && strings.HasSuffix(entry.Name(), ".json") {
-			// Parse filename: serviceName_serviceID.json
-			name := strings.TrimSuffix(entry.Name(), ".json")
-			parts := strings.SplitN(name, "_", 2)
-			if len(parts) != 2 {
-				continue
-			}
-
-			serviceName := parts[0]
-			serviceID := parts[1]
-
-			// Load global services (non-etcd services can be included here)
-			if serviceName == "multiadmin" || serviceName == "etcd" {
-				req := &provisioner.DeprovisionRequest{
-					Service:      serviceName,
-					ServiceID:    serviceID,
-					DatabaseName: "", // global services have no database name
-				}
-
-				service, err := p.loadServiceState(req)
-				if err != nil {
-					// Log warning but continue with other services
-					fmt.Printf("Warning: failed to load state for global service %s: %v\n", serviceID, err)
-					continue
-				}
-
-				if service != nil {
-					services = append(services, service)
-				}
-			}
-		}
-	}
-
-	return services, nil
-}
-
-// findRunningDbService finds a running service by service name within a specific database and cell
-func (p *localProvisioner) findRunningDbService(serviceName, databaseName, cell string) (*LocalProvisionedService, error) {
-	services, err := p.loadDbProvisionedServices(databaseName)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load service states for database %s: %w", databaseName, err)
-	}
-
-	for _, service := range services {
-		if service.Service == serviceName {
-			// Check if the service matches the cell
-			if serviceCell, ok := service.Metadata["cell"].(string); ok && serviceCell == cell {
-				// Check if the service is actually still running
-				if service.PID > 0 {
-					if err := p.validateProcessRunning(service.PID); err == nil {
-						return service, nil
-					}
-				}
-			}
-		}
-	}
-
-	// Check for port conflicts with other processes using cell-specific config
-	expectedPorts := p.getExpectedPortsForDbService(serviceName, cell)
-	for portName, port := range expectedPorts {
-		if err := p.checkPortConflict(port, serviceName, portName); err != nil {
-			return nil, err
-		}
-	}
-
-	return nil, nil // No running service found
-}
-
-// getExpectedPortsForDbService returns expected ports for a DB-scoped service (per cell)
-func (p *localProvisioner) getExpectedPortsForDbService(serviceName, cell string) map[string]int {
-	ports := make(map[string]int)
-
-	cellConfig, err := p.getCellServiceConfig(cell, serviceName)
-	if err != nil {
-		return ports
-	}
-
-	switch serviceName {
-	case "multigateway":
-		if httpPort, ok := cellConfig["http_port"].(int); ok {
-			ports["http"] = httpPort
-		}
-		if grpcPort, ok := cellConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-	case "multipooler":
-		if grpcPort, ok := cellConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-		if httpPort, ok := cellConfig["http_port"].(int); ok && httpPort > 0 {
-			ports["http"] = httpPort
-		}
-	case "multiorch":
-		if grpcPort, ok := cellConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-		if httpPort, ok := cellConfig["http_port"].(int); ok && httpPort > 0 {
-			ports["http"] = httpPort
-		}
-	}
-
-	return ports
-}
-
-// getExpectedPortsForService returns the expected ports for a service based on its configuration
-func (p *localProvisioner) getExpectedPortsForService(serviceName string) map[string]int {
-	serviceConfig := p.getServiceConfig(serviceName)
-	ports := make(map[string]int)
-
-	switch serviceName {
-	case "multigateway":
-		if httpPort, ok := serviceConfig["http_port"].(int); ok {
-			ports["http"] = httpPort
-		}
-		if grpcPort, ok := serviceConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-	case "multipooler":
-		if grpcPort, ok := serviceConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-		if httpPort, ok := serviceConfig["http_port"].(int); ok && httpPort > 0 {
-			ports["http"] = httpPort
-		}
-	case "multiorch":
-		if grpcPort, ok := serviceConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-		if httpPort, ok := serviceConfig["http_port"].(int); ok && httpPort > 0 {
-			ports["http"] = httpPort
-		}
-	case "multiadmin":
-		if httpPort, ok := serviceConfig["http_port"].(int); ok {
-			ports["http"] = httpPort
-		}
-		if grpcPort, ok := serviceConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-	case "etcd":
-		if port, ok := serviceConfig["port"].(int); ok {
-			ports["tcp"] = port
-		}
-	}
-
-	return ports
-}
-
-// checkPortConflict checks if a port is already in use by another process
-func (p *localProvisioner) checkPortConflict(port int, serviceName, portName string) error {
-	if port <= 0 {
-		return nil // Skip invalid ports
-	}
-
-	address := fmt.Sprintf("localhost:%d", port)
-	conn, err := net.DialTimeout("tcp", address, 1*time.Second)
-	if err != nil {
-		// Port is not in use, this is good
-		return nil
-	}
-	conn.Close()
-
-	// Port is in use by some process
-	return fmt.Errorf("cannot start %s: port %d (%s) is already in use by another process. "+
-		"There is no way to do a clean start. Please kill the process using port %d or change the configuration",
-		serviceName, port, portName, port)
-}
-
-// findRunningEtcdService finds a running etcd service
-func (p *localProvisioner) findRunningEtcdService() (*LocalProvisionedService, error) {
-	services, err := p.loadEtcdServices()
-	if err != nil {
-		return nil, fmt.Errorf("failed to load etcd service states: %w", err)
-	}
-
-	for _, service := range services {
-		if service.Service == "etcd" {
-			// Check if the service is actually still running
-			if service.PID > 0 {
-				if err := p.validateProcessRunning(service.PID); err == nil {
-					return service, nil
-				}
-			}
-		}
-	}
-
-	// Check for port conflicts with other processes
-	expectedPorts := p.getExpectedPortsForService("etcd")
-	for portName, port := range expectedPorts {
-		if err := p.checkPortConflict(port, "etcd", portName); err != nil {
-			return nil, err
-		}
-	}
-
-	return nil, nil // No running etcd service found
-}
-
-// findRunningService finds a running service by service name (for global services like multiadmin)
-func (p *localProvisioner) findRunningService(serviceName string) (*LocalProvisionedService, error) {
-	// Load global services (e.g., multiadmin, etcd)
-	services, err := p.loadGlobalServices()
-	if err != nil {
-		return nil, fmt.Errorf("failed to load global service states: %w", err)
-	}
-
-	for _, service := range services {
-		if service.Service == serviceName {
-			// Check if the service is actually still running
-			if service.PID > 0 {
-				if err := p.validateProcessRunning(service.PID); err == nil {
-					return service, nil
-				}
-			}
-		}
-	}
-
-	// Check for port conflicts with other processes
-	expectedPorts := p.getExpectedPortsForService(serviceName)
-	for portName, port := range expectedPorts {
-		if err := p.checkPortConflict(port, serviceName, portName); err != nil {
-			return nil, err
-		}
-	}
-
-	return nil, nil // No running service found
-}
-
 // stopService stops a specific service based on its type using the internal methods
 func (p *localProvisioner) stopService(ctx context.Context, req *provisioner.DeprovisionRequest) error {
 	switch req.Service {
@@ -1832,42 +961,6 @@ func (p *localProvisioner) deprovisionService(ctx context.Context, req *provisio
 		}
 	}
 
-	return nil
-}
-
-// stopProcessByPID stops a process by its PID
-func (p *localProvisioner) stopProcessByPID(pid int) error {
-	// Check if process exists
-	process, err := os.FindProcess(pid)
-	if err != nil {
-		// Process not found, assume already cleaned up
-		fmt.Printf("Process %d not found, assuming already stopped\n", pid)
-		return nil
-	}
-
-	// Send SIGTERM to gracefully stop the process
-	if err := process.Signal(syscall.SIGTERM); err != nil {
-		// Process might already be dead, check errno
-		if err.Error() == "no such process" || err.Error() == "process already finished" {
-			fmt.Printf("Process %d already stopped\n", pid)
-			return nil
-		}
-
-		// If SIGTERM fails for other reasons, try SIGKILL
-		if err := process.Kill(); err != nil {
-			// If kill also fails and it's because process doesn't exist, that's ok
-			if err.Error() == "no such process" || err.Error() == "process already finished" {
-				fmt.Printf("Process %d already stopped\n", pid)
-				return nil
-			}
-			return fmt.Errorf("failed to kill process %d: %w", pid, err)
-		}
-	}
-
-	// Wait a bit for the process to exit
-	time.Sleep(2 * time.Second)
-
-	fmt.Printf("Process %d stopped successfully\n", pid)
 	return nil
 }
